requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
pydantic==2.7.4
python-dotenv==1.0.1
psycopg2-binary==2.9.11
openai==1.35.13
google-api-python-client==2.132.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.2.2         # MODIFIED for guaranteed stability and wheel availability on Python 3.11
python-dateutil==2.9.0.post0